+++ R/summary.R
@@ -4,7 +4,9 @@ summary.estimate_slopes <- function(object, verbose = TRUE, ...) {
   by <- attributes(object)$by
 
   if (verbose && nrow(out) < 50) {
-    insight::format_alert("There might be too few data to accurately determine intervals. Consider setting `length = 100` (or larger) in your call to `estimate_slopes()`.") # nolint
+    insight::format_alert(
+      "There might be too few data to accurately determine intervals. Consider setting `length = 100` (or larger) in your call to `estimate_slopes()`."
+    ) # nolint
   }
 
   # Add "Confidence" col based on the sig index present in the data
@@ -20,6 +22,10 @@ summary.estimate_slopes <- function(object, verbose = TRUE, ...) {
     out$Group <- gsub("\\.\\d+$", "", out$Group)
   } else {
     out <- .estimate_slope_parts(out, by)
+  }
+
+  if (anyNA(out$Confidence) && verbose) {
+    insight::format_warning("Significance could not be determined for some intervals.")
   }
 
   attributes(out) <- utils::modifyList(attributes(object), attributes(out))
@@ -40,7 +46,6 @@ summary.reshape_grouplevel <- function(object, ...) {
 
 # Utilities ===============================================================
 
-
 .estimate_slope_parts <- function(out, by) {
   # mark all "changes" from negative to positive and vice versa
   index <- 1
@@ -48,7 +53,16 @@ summary.reshape_grouplevel <- function(object, ...) {
   index <- index + 1
 
   for (i in 2:nrow(out)) {
-    if (out$Direction[i] != out$Direction[i - 1] || out$Confidence[i] != out$Confidence[i - 1]) {
+    # we only proceed if we have non-missing values
+    do_switch <- (!is.na(out$Direction[i - 1]) & !is.na(out$Direction[i])) &&
+      out$Direction[i] != out$Direction[i - 1]
+
+    if (!do_switch) {
+      do_switch <- (!is.na(out$Confidence[i - 1]) & !is.na(out$Confidence[i])) &&
+        out$Confidence[i] != out$Confidence[i - 1]
+    }
+
+    if (do_switch) {
       out$switch[i:nrow(out)] <- index # styler: off
       index <- index + 1
     }
@@ -57,14 +71,17 @@ summary.reshape_grouplevel <- function(object, ...) {
   # split into "switches"
   parts <- split(out, out$switch)
 
-  do.call(rbind, lapply(parts, function(i) {
-    data.frame(
-      Start = i[[by]][1],
-      End = i[[by]][nrow(i)],
-      Direction = i$Direction[1],
-      Confidence = i$Confidence[1]
-    )
-  }))
+  do.call(
+    rbind,
+    lapply(parts, function(i) {
+      data.frame(
+        Start = i[[by]][1],
+        End = i[[by]][nrow(i)],
+        Direction = i$Direction[1],
+        Confidence = i$Confidence[1]
+      )
+    })
+  )
 }
 
 
@@ -83,18 +100,27 @@ summary.reshape_grouplevel <- function(object, ...) {
 
   if (confidence == "auto") {
     # TODO: make sure all of these work
-    if ("BF" %in% names(x)) confidence <- "BF"
-    if ("p" %in% names(x)) confidence <- "p"
+    if ("BF" %in% names(x)) {
+      confidence <- "BF"
+    }
+    if ("p" %in% names(x)) {
+      confidence <- "p"
+    }
     if ("pd" %in% names(x)) confidence <- "pd"
   }
 
-  switch(confidence,
+  switch(
+    confidence,
     p = tools::toTitleCase(effectsize::interpret_p(x$p, ...)),
     BF = tools::toTitleCase(effectsize::interpret_bf(x$BF, ...)),
     pd = tools::toTitleCase(effectsize::interpret_pd(x$pd, ...)),
     {
       # Based on CI
-      out <- ifelse((x$CI_high < 0 & x$CI_low < 0) | (x$CI_high > 0 & x$CI_low > 0), "Significant", "Uncertain")
+      out <- ifelse(
+        (x$CI_high < 0 & x$CI_low < 0) | (x$CI_high > 0 & x$CI_low > 0),
+        "Significant",
+        "Uncertain"
+      )
       factor(out, levels = c("Uncertain", "Significant"))
     }
   )